Cambios en la forma de traer publicaciones (Se creo un DTO para traer publicaciones con el nombre de quien las creo, Ademas se crearon consultas nativas en el repositorio de publicacion para traer las publicaciones dependiendo del rol de quien la creo), Se crearon consultas nativas para traer mensajes segun el id de cuenta ( Esto es para que se muestren los mensajes que se envian a una cuenta especifica)

+++ src/main/java/com/example/SAC/repository/AreaComunRepository.java
@@ -5,5 +5,6 @@ import com.example.SAC.entity.AreaComun;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AreaComunRepository extends JpaRepository<AreaComun, Integer> {
+public interface AreaComunRepository extends JpaRepository<AreaComun, Long> {
+
 }
